Extract Parser test cases into each file
